+++ lifecycle/lifecycle/endpoints/auth.py
@@ -70,9 +70,9 @@ def setup_auth_endpoints(api: APIRouter, config: Config):
     
     class JobCallAuthData(BaseModel):
         job: JobDto
-        caller: str | None
-        remote_gateway_url: str | None  # URL to remote PUB gateway
-        remote_gateway_token: str | None
+        caller: str | None = None
+        remote_gateway_url: str | None = None  # URL to remote PUB gateway
+        remote_gateway_token: str | None = None
 
     @api.get('/auth/can-call-job/{job_name}/{job_version}/{endpoint:path}', response_model=JobCallAuthData)
     def _auth_can_call_job_endpoint(