+++ lifecycle/lifecycle/endpoints/deploy.py
@@ -1,6 +1,8 @@
 from typing import Any, Dict, Optional
 from fastapi import APIRouter, Request
 
+from pydantic import BaseModel, ConfigDict, Field
+
 from lifecycle.auth.check import check_auth
 from lifecycle.auth.authenticate import get_username_from_token
 from lifecycle.config import Config
@@ -9,7 +11,6 @@ from lifecycle.deployer.builder import build_job_in_background
 from lifecycle.deployer.deploy import deploy_job_in_background
 from lifecycle.job.deployment import check_deployment_result, save_deployment_phase, list_recent_deployments
 from lifecycle.server.metrics import metric_requested_job_deployments
-from pydantic import BaseModel, Field
 from racetrack_commons.plugin.engine import PluginEngine
 from racetrack_client.client.env import load_secret_vars_from_dict
 from racetrack_client.client_config.io import load_credentials_from_dict
@@ -18,13 +19,15 @@ from racetrack_client.manifest.load import load_manifest_from_dict
 
 def setup_deploy_endpoints(api: APIRouter, config: Config, plugin_engine: PluginEngine):
     class CredentialsModel(BaseModel):
-        username: str = Field(example="admin")
-        password: str = Field(example="hunter2")
+        username: str = Field(examples=["admin"])
+        password: str = Field(examples=["hunter2"])
 
-    class DeployPayloadModel(BaseModel, arbitrary_types_allowed=True):
+    class DeployPayloadModel(BaseModel):
+        model_config = ConfigDict(arbitrary_types_allowed=True)
+
         manifest: Dict[str, Any] = Field(
             description='Manifest - build recipe for a Job',
-            example={
+            examples=[{
                 'name': 'adder',
                 'jobtype': 'python3',
                 'owner_email': 'nobody@example.com',
@@ -36,7 +39,7 @@ def setup_deploy_endpoints(api: APIRouter, config: Config, plugin_engine: Plugin
                     'requirements_path': 'requirements.txt',
                     'entrypoint_path': 'adder.py',
                 },
-            },
+            }],
         )
         git_credentials: Optional[CredentialsModel] = Field(
             default=None,
@@ -45,24 +48,24 @@ def setup_deploy_endpoints(api: APIRouter, config: Config, plugin_engine: Plugin
         secret_vars: Optional[Dict[str, Any]] = Field(
             default=None,
             description='secret environment vars',
-            example={
+            examples=[{
                 'build_env': {
                     'GIT_PASS': '5ecr3t',
                 },
                 'runtime_env': {
                     'PGPASSWORD': '5ecr3t',
                 },
-            },
+            }],
         )
         build_context: Optional[str] = Field(
             default=None,
             description='encoded build context',
-            example='',
+            examples=[''],
         )
         force: Optional[bool] = Field(
             default=None,
             description='overwrite existing job',
-            example=False,
+            examples=[False],
         )
 
     @api.post('/deploy')