+++ lifecycle/lifecycle/infrastructure/model.py
@@ -1,10 +1,12 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 
 from lifecycle.deployer.base import JobDeployer
 from lifecycle.monitor.base import JobMonitor, LogsStreamer
 
 
-class InfrastructureTarget(BaseModel, arbitrary_types_allowed=True):
+class InfrastructureTarget(BaseModel):
+    model_config = ConfigDict(arbitrary_types_allowed=True)
+
     name: str | None = None
     job_deployer: JobDeployer | None = None
     job_monitor: JobMonitor | None = None