+++ docs/CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Take a look at the [python-logging-format sample](../sample/python-logging-format/entrypoint.py)
   to see how to configure your own logging formatter in jobs.
 
+### Changed
+- `pydantic` package has been upgraded to version 2 to prevent from conflicts with other up-to-date packages.
+  `pydantic` is one of the dependencies of the `racetrack-client` package.
+
 ## [2.26.0] - 2024-01-24
 ### Changed
 - Job type plugins can now build job images from a single Dockerfile.